feat: implement analytics service and dashboard with summary and category breakdown
- Add analytics service with monthly summary and category breakdown
- Implement analytics API endpoints with Redis caching (1 hour TTL)
- Create Dashboard component with summary cards and category visualization
- Add comprehensive tests for analytics service and API endpoints
- All tests passing (7/7)

+++ backend/app/schemas/analytics.py
@@ -16,9 +16,9 @@ class TopCategory(BaseModel):
 class SummaryResponse(BaseModel):
     """Schema for financial summary response"""
     month: str
-    total_income: Decimal
-    total_expense: Decimal
-    balance: Decimal
+    total_income: float
+    total_expense: float
+    balance: float
     transaction_count: int
     top_category: Optional[TopCategory] = None
 
@@ -26,7 +26,7 @@ class SummaryResponse(BaseModel):
 class CategoryBreakdown(BaseModel):
     """Schema for category breakdown"""
     name: str
-    amount: Decimal
+    amount: float
     percentage: float
     count: int
